backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.9.2
pydantic-settings==2.5.2
better-exceptions==0.3.3
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==43.0.1
uuid==1.30